+++ backend/app/Http/Controllers/PodcastController.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Requests\UpdatePodcastRequest;
-use App\Http\Requests\StorePodcastRequest;   
+use App\Http\Requests\StorePodcastRequest;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
@@ -28,8 +28,8 @@ class PodcastController extends Controller
     public function index()
     {
         $podkasts = Podcast::join('user_metadata', 'podcasts.author_id', '=', 'user_metadata.user_id')
-                ->select('podcasts.*', 'user_metadata.first_name', 'user_metadata.last_name', 'user_metadata.patronymic', 'user_metadata.nickname')
-                ->get();
+            ->select('podcasts.*', 'user_metadata.first_name', 'user_metadata.last_name', 'user_metadata.patronymic', 'user_metadata.nickname')
+            ->get();
         return response()->json($podkasts);
     }
 
@@ -47,19 +47,24 @@ class PodcastController extends Controller
      * @bodyParam podcastId int ID подкаста.
      * @urlParam withAuthors bool Включать авторов в ответ.
      * @urlParam page int Номер страницы.
+     * @urlParam searchFields string[] Массив столбцов для поиска.
+     * @urlParam searchValues string[] Массив значений для поиска.
      * @urlParam searchColumnName string Поиск по столбцу.
      * @urlParam searchValue string Поисковый запрос.
      * @urlParam tagFilter string Фильтр по тегу в meta описания.
-     * @urlParam crtFrom string Дата начала (формат: Y-m-d H:i:s).
-     * @urlParam crtTo string Дата окончания (формат: Y-m-d H:i:s).
-     * @urlParam updFrom string Дата начала (формат: Y-m-d H:i:s).
-     * @urlParam updTo string Дата окончания (формат: Y-m-d H:i:s).
+     * @urlParam crtFrom string Дата начала (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam crtTo string Дата окончания (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam crtDate string Дата создания (формат: Y-m-d).
+     * @urlParam updFrom string Дата начала (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam updTo string Дата окончания (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam updDate string Дата обновления (формат: Y-m-d).
+     * @urlParam operator string Логический оператор для условий поиска ('and' или 'or').
      * 
      * @param \Illuminate\Http\Request $request
      * @return mixed|\Illuminate\Http\JsonResponse
      */
     public function getPodcasts(Request $request)
-    {
+    {//TODO: Переделать
         if (!Auth::user()->can('view', Podcast::class)) {
             return $this->errorResponse('Нет прав на просмотр', [], 403);
         }
@@ -69,6 +74,8 @@ class PodcastController extends Controller
         $currentUser = $request->query('currentUser');
         $podcastId = $request->query('podcastId');
         $withAuthors = $request->query('withAuthors', false);
+        $searchFields = $request->query('searchFields', []);
+        $searchValues = $request->query('searchValues', []);
         $searchColumnName = $request->query('searchColumnName');
         $searchValue = $request->query('searchValue');
         $tagFilter = $request->query('tagFilter');
@@ -76,6 +83,11 @@ class PodcastController extends Controller
         $crtTo = $request->query('crtTo');
         $updFrom = $request->query('updFrom');
         $updTo = $request->query('updTo');
+
+        $updDate = $request->query('updDate');
+        $crtDate = $request->query('crtDate');
+
+        $operator = $request->query('operator', 'and');
 
         $query = Podcast::query();
 
@@ -99,6 +111,31 @@ class PodcastController extends Controller
             }
         } elseif ($podcastId) {
             $query->where('id', $podcastId);
+            $podcast = $query->first(); // Получаем первый подкаст, соответствующий запросу
+            if ($podcast) {
+                $podcast->increment('views'); // Инкрементируем счетчик просмотров
+            }
+        }
+
+
+        if (!empty($searchFields) && !empty($searchValues)) {
+            if ($operator === 'or') {
+                $query->where(function ($query) use ($searchFields, $searchValues) {
+                    foreach ($searchFields as $index => $field) {
+                        $value = $searchValues[$index] ?? null;
+                        if ($value) {
+                            $query->orWhere($field, 'LIKE', '%' . $value . '%');
+                        }
+                    }
+                });
+            } else {
+                foreach ($searchFields as $index => $field) {
+                    $value = $searchValues[$index] ?? null;
+                    if ($value) {
+                        $query->where($field, 'LIKE', '%' . $value . '%');
+                    }
+                }
+            }
         }
 
         if ($searchColumnName) {
@@ -108,6 +145,11 @@ class PodcastController extends Controller
         if ($tagFilter) {
             $query->whereRaw("description->'meta'->>'tags' LIKE ?", ['%' . $tagFilter . '%']);
         }
+
+        $crtFrom = $this->parseDate($crtFrom);
+        $crtTo = $this->parseDate($crtTo);
+        $updFrom = $this->parseDate($updFrom);
+        $updTo = $this->parseDate($updTo);
 
         if ($crtFrom && $crtTo) {
             $query->whereBetween('created_at', [$crtFrom, $crtTo]);
@@ -125,6 +167,14 @@ class PodcastController extends Controller
             $query->where('updated_at', '<=', $updTo);
         }
 
+        if ($crtDate) {
+            $query->whereDate('created_at', '=', $crtDate);
+        }
+
+        if ($updDate) {
+            $query->whereDate('updated_at', '=', $updDate);
+        }
+
         $podcasts = $query->paginate($perPage);
 
         $paginationData = [
@@ -139,7 +189,25 @@ class PodcastController extends Controller
         return $this->successResponse($podcasts->items(), $paginationData, 200);
     }
 
+    /**
+     * Parses the date from the given input.
+     * Supports both Y-m-d H:i:s and Y-m-d formats.
+     * 
+     * @param string|null $date
+     * @return string|null
+     */
+    private function parseDate($date)
+    {
+        if (!$date) {
+            return null;
+        }
 
+        if (preg_match('/^\d{4}-\d{2}-\d{2}$/', $date)) {
+            return $date . ' 00:00:00';
+        }
+
+        return $date;
+    }
 
 
     /**
@@ -158,31 +226,19 @@ class PodcastController extends Controller
      */
     public function store(StorePodcastRequest $request)
     {
-        try {
-            // Проверка прав пользователя
-            if (!Auth::user()->can('create', Podcast::class)) {
-                throw new AccessDeniedHttpException('You do not have permission to create a podcast');
-            }
-
-            $this->validateRequest($request, $request->rules());
-
-            // Создание нового подкаста с использованием проверенных данных
-            $podcast = Podcast::create(array_merge($request->validated(), [
-                'status' => 'moderating',
-                'author_id' => Auth::id(),
-            ]));
-            return $this->successResponse(['podcast' => $podcast], 'Podcast created successfully', 200);
-        } catch (AccessDeniedHttpException $e) {
-            // Обработка исключений через централизованную функцию
-            return $this->handleException($e);
+        if (!Auth::user()->can('create', Podcast::class)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
         }
+
+        $podcast = Podcast::create($request->validated() + [
+            'status' => 'moderating',
+            'author_id' => Auth::id(),
+        ]);     
+
+        return $this->successResponse(['podcast' => $podcast], 'Podcast created successfully', 200);
     }
 
-
-  
-
-
-   /**
+    /**
      * Обновить
      * 
      * Обновление подкаста
@@ -200,27 +256,52 @@ class PodcastController extends Controller
      */
     public function update(UpdatePodcastRequest $request, $id)
     {
-        try{
+        $podcast = Podcast::find($id);
 
-            $podcast = Podcast::findOrFail($id);
-
-            if(!Auth::user()->can('update', $podcast)) {
-                throw new AccessDeniedHttpException('You do not have permission to update this podcast');
-            }
-
-            $this->validateRequest($request, $request->rules());
-            $validatedData = $request->validated();
-            $podcast->update($validatedData);
-
-            return $this->successResponse(['podcast' => $podcast], 'Podcast updated successfully', 200);
-        } catch (AccessDeniedHttpException $e) {
-            return $this->handleException($e);
-        } catch (ModelNotFoundException $e) {
-            return $this->handleException($e);
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
+
+        if (!Auth::user()->can('update', $podcast)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        $validatedData = $request->validated();
+        $podcast->update($validatedData);
+
+        return $this->successResponse(['podcast' => $podcast], 'Podcast updated successfully', 200);
     }
 
+    /**
+     * Обновить
+     * 
+     * Обновление статуса подкаста
+     * 
+     * @group Подкасты
+     * 
+     * @bodyParam status string required Статус
+     */
+    public function updateStatus(int $id, Request $request): \Illuminate\Http\JsonResponse
+    {
+        $newStatus = $request->input('status');
+        $podcast = Podcast::find($id);
 
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
+        }
+
+        if (!Auth::user()->can('updateStatus', $podcast)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        if (!in_array($newStatus, Podcast::STATUSES)) {
+            return $this->errorResponse('Invalid status entered', [], 404);
+        }
+
+        $podcast->update(['status' => $newStatus]);
+
+        return $this->successResponse(['podcasts' => $podcast], 'Podcast status updated successfully', 200);
+    }
 
 
     /**
@@ -235,21 +316,80 @@ class PodcastController extends Controller
      */
     public function destroy($id)
     {
-        try{
-            $podcast = Podcast::findOrFail($id);
+        $podcast = Podcast::find($id);
 
-            // Проверка прав пользователя
-            if (!Auth::user()->can('delete', $podcast)) {
-                throw new AccessDeniedHttpException('You do not have permission to delete this podcast');
-            }
-
-            $podcast->delete();
-
-            return $this->successResponse(['podcast' => $podcast], 'Podcast deleted successfully', 200);
-        }catch (Exception $e) {
-            return $this->handleException($e);
-        } catch (ModelNotFoundException $e) {
-            return $this->handleException($e);
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
+
+        // Проверка прав пользователя
+        if (!Auth::user()->can('delete', $podcast)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        $podcast->delete();
+
+        return $this->successResponse(['podcast' => $podcast], 'Podcast deleted successfully', 200);
+    }
+
+    /**
+     * Лайкнуть подкаст
+     * 
+     * Этот метод позволяет пользователю "лайкнуть" или "дизлайкнуть" подкаст.
+     * 
+     * @group    Подкасты
+     * 
+     * @urlParam id int Обязательно. Идентификатор подкаста.
+     * 
+     * @response 200 {
+     *   "podcasts": {
+     *     "id": 1,
+     *     "title": "Название подкаста",
+     *     "description": "Описание подкаста",
+     *     "content": "Содержание подкаста",
+     *     "cover_uri": "URI обложки подкаста",
+     *     "status": "Статус подкаста",
+     *     "views": 100,
+     *     "likes": 50,
+     *     "reposts": 20
+     *   },
+     *   "message": "Podcast liked successfully"
+     * }
+     * 
+     * @response 200 {
+     *   "podcasts": {
+     *     "id": 1,
+     *     "title": "Название подкаста",
+     *     "description": "Описание подкаста",
+     *     "content": "Содержание подкаста",
+     *     "cover_uri": "URI обложки подкаста",
+     *     "status": "Статус подкаста",
+     *     "views": 100,
+     *     "likes": 49,
+     *     "reposts": 20
+     *   },
+     *   "message": "Podcast unliked successfully"
+     * }
+     */
+    public function likePodcast(Request $request, int $id): \Illuminate\Http\JsonResponse
+    {
+        $podcast = Podcast::find($id);
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
+        }
+        $user = Auth::user();
+
+        $like = $podcast->likes()->where('user_id', $user->id)->first();
+
+        if ($like) {
+            $like->delete();
+            $podcast->decrement('likes');
+            return $this->successResponse(['podcasts' => $podcast], 'Podcast unliked successfully', 200);
+        } else {
+            $podcast->likes()->create(['user_id' => $user->id]);
+            $podcast->increment('likes');
+        }
+
+        return $this->successResponse(['podcasts' => $podcast], 'Podcast liked successfully', 200);
     }
 }
