+++ backend/app/Http/Controllers/ProjectController.php
@@ -30,19 +30,24 @@ class ProjectController extends Controller
      * @bodyParam projectId int ID проекта.
      * @urlParam withAuthors bool Включать авторов в ответ.
      * @urlParam page int Номер страницы.
+     * @urlParam searchFields string[] Массив столбцов для поиска.
+     * @urlParam searchValues string[] Массив значений для поиска.
      * @urlParam searchColumnName string Поиск по столбцу.
      * @urlParam searchValue string Поисковый запрос.
      * @urlParam tagFilter string Фильтр по тегу в meta описания.
-     * @urlParam crtFrom string Дата начала (формат: Y-m-d H:i:s).
-     * @urlParam crtTo string Дата окончания (формат: Y-m-d H:i:s).
-     * @urlParam updFrom string Дата начала (формат: Y-m-d H:i:s).
-     * @urlParam updTo string Дата окончания (формат: Y-m-d H:i:s).
+     * @urlParam crtFrom string Дата начала (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam crtTo string Дата окончания (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam crtDate string Дата создания (формат: Y-m-d).
+     * @urlParam updFrom string Дата начала (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam updTo string Дата окончания (формат: Y-m-d H:i:s или Y-m-d).
+     * @urlParam updDate string Дата обновления (формат: Y-m-d).
+     * @urlParam operator string Логический оператор для условий поиска ('and' или 'or').
      * 
      * @param \Illuminate\Http\Request $request
      * @return mixed|\Illuminate\Http\JsonResponse
      */
     public function getProjects(Request $request)
-    {
+    {//TODO: Переделать
         if (!Auth::user()->can('view', Project::class)) {
             return $this->errorResponse('Нет прав на просмотр', [], 403);
         }
@@ -54,11 +59,18 @@ class ProjectController extends Controller
         $withAuthors = $request->query('withAuthors', false);
         $searchColumnName = $request->query('searchColumnName');
         $searchValue = $request->query('searchValue');
+        $searchFields = $request->query('searchFields', []);
+        $searchValues = $request->query('searchValues', []);
         $tagFilter = $request->query('tagFilter');
         $crtFrom = $request->query('crtFrom');
         $crtTo = $request->query('crtTo');
         $updFrom = $request->query('updFrom');
         $updTo = $request->query('updTo');
+
+        $updDate = $request->query('updDate');
+        $crtDate = $request->query('crtDate');
+
+        $operator = $request->query('operator', 'and');
 
         $query = Project::query();
 
@@ -84,6 +96,26 @@ class ProjectController extends Controller
             $query->where('id', $projectId);
         }
 
+        if (!empty($searchFields) && !empty($searchValues)) {
+            if ($operator === 'or') {
+                $query->where(function ($query) use ($searchFields, $searchValues) {
+                    foreach ($searchFields as $index => $field) {
+                        $value = $searchValues[$index] ?? null;
+                        if ($value) {
+                            $query->orWhere($field, 'LIKE', '%' . $value . '%');
+                        }
+                    }
+                });
+            } else {
+                foreach ($searchFields as $index => $field) {
+                    $value = $searchValues[$index] ?? null;
+                    if ($value) {
+                        $query->where($field, 'LIKE', '%' . $value . '%');
+                    }
+                }
+            }
+        }
+
         if ($searchColumnName) {
             $query->where($searchColumnName, 'LIKE', '%' . $searchValue . '%');
         }
@@ -91,6 +123,11 @@ class ProjectController extends Controller
         if ($tagFilter) {
             $query->whereRaw("description->'meta'->>'tags' LIKE ?", ['%' . $tagFilter . '%']);
         }
+
+        $crtFrom = $this->parseDate($crtFrom);
+        $crtTo = $this->parseDate($crtTo);
+        $updFrom = $this->parseDate($updFrom);
+        $updTo = $this->parseDate($updTo);
 
         if ($crtFrom && $crtTo) {
             $query->whereBetween('created_at', [$crtFrom, $crtTo]);
@@ -108,6 +145,14 @@ class ProjectController extends Controller
             $query->where('updated_at', '<=', $updTo);
         }
 
+        if ($crtDate) {
+            $query->whereDate('created_at', '=', $crtDate);
+        }
+    
+        if ($updDate) {
+            $query->whereDate('updated_at', '=', $updDate);
+        }
+
         $projects = $query->paginate($perPage);
 
         $paginationData = [
@@ -122,39 +167,48 @@ class ProjectController extends Controller
         return $this->successResponse($projects->items(), $paginationData, 200);
     }
 
+    /**
+     * Parses the date from the given input.
+     * Supports both Y-m-d H:i:s and Y-m-d formats.
+     * 
+     * @param string|null $date
+     * @return string|null
+     */
+    private function parseDate($date)
+    {
+        if (!$date) {
+            return null;
+        }
 
+        if (preg_match('/^\d{4}-\d{2}-\d{2}$/', $date)) {
+            return $date . ' 00:00:00';
+        }
+
+        return $date;
+    }
 
 
     /**
      * Store a newly created resource in storage.
+     *
+     * @param \App\Http\Requests\StoreProjectRequest $request The request object containing the project data.
+     * @return \Illuminate\Http\JsonResponse The JSON response containing the created project.
+     *
+     * @throws \Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException If the user does not have permission to create a project.
      */
-    public function store(StoreProjectRequest $request)
+    public function store(StoreProjectRequest $request): \Illuminate\Http\JsonResponse
     {
-        try {
-            if (!Auth::user()->can('create', Project::class)) {
-                throw new AccessDeniedHttpException('You do not have permission to create a project');
-            }
-
-            $this->validateRequest($request, $request->rules());
-
-            $project = Project::create(array_merge($request->validated(), [
-                'author_id' => Auth::id(),
-            ]));
-
-            return $this->successResponse(['projects' => $project], 'Project created successfully', 231);
-        } catch (AccessDeniedHttpException $e) {
-            return $this->handleException($e);
+        if (!Auth::user()->can('create', Project::class)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
         }
+
+        $project = Project::create($request->validated() + [
+            'author_id' => Auth::id(),
+        ]);
+
+        return $this->successResponse(['projects' => $project], 'Project created successfully', 231);
     }
 
-
-
-
-
-    public function storeTHIS(StoreProjectRequest $request)
-    {
-        Log::info('checkpoint');
-    }
 
 
 
@@ -162,29 +216,29 @@ class ProjectController extends Controller
 
     /**
      * Update the specified resource in storage.
+     *
+     * @param \App\Http\Requests\UpdateProjectRequest $request The request object containing the updated project data.
+     * @param int $id The ID of the project to update.
+     * @return \Illuminate\Http\JsonResponse The JSON response containing the updated project.
+     * @throws \Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException If the user does not have permission to update the project.
+     * @throws \Illuminate\Database\Eloquent\ModelNotFoundException If the project with the given ID is not found.
      */
-    public function update(UpdateProjectRequest $request, int $id)
+    public function update(UpdateProjectRequest $request, int $id): \Illuminate\Http\JsonResponse
     {
-        try {
-            $project = Project::findOrFail($id);
+        $project = Project::find($id);
 
-            if (!Auth::user()->can('update', $project)) {
-                throw new AccessDeniedHttpException('You do not have permission to update this project');
-            }
-
-            $project->update($request->validated());
-
-            return $this->successResponse(['projects' => $project], 'Project updated successfully', 200);
-        } catch (AccessDeniedHttpException $e) {
-            return $this->handleException($e);
-        } catch (ModelNotFoundException $e) {
-            Log::info('catch_error', [$e]);
-            return $this->handleException($e);
+        if (!$project) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
+
+        if (!Auth::user()->can('update', $project)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        $project->update($request->validated());
+
+        return $this->successResponse(['projects' => $project], 'Project updated successfully', 200);
     }
-
-
-
 
     /**
      * Remove the specified resource from storage.
@@ -196,22 +250,18 @@ class ProjectController extends Controller
      */
     public function destroy(int $id): \Illuminate\Http\JsonResponse
     {
-        try {
-            $project = Project::findOrFail($id);
+        $project = Project::find($id);
 
-            if (!Auth::user()->can('delete', $project)) {
-                throw new AccessDeniedHttpException('You do not have permission to delete this project');
-            }
-
-            $project->delete();
-
-            return $this->successResponse(['projects' => $project], 'Project deleted successfully', 200);
-        } catch (AccessDeniedHttpException $e) {
-            Log::info('catch_error', [$e]);
-            return $this->handleException($e);
-        } catch (ModelNotFoundException $e) {
-            Log::info('catch_error', [$e]);
-            return $this->handleException($e);
+        if (!$project) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
+
+        if (!Auth::user()->can('delete', $project)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        $project->delete();
+
+        return $this->successResponse(['projects' => $project], 'Project deleted successfully', 200); 
     }
 }
